+++ packages/dso-toolkit/src/components/info-button/info-button.models.ts
@@ -1,8 +1,9 @@
-export interface InfoButton {
+export interface InfoButton<TemplateFnReturnType> {
   active?: boolean;
   secondary?: boolean;
   label?: string;
   dsoToggle?: (e: CustomEvent<InfoButtonToggleEvent>) => void;
+  children?: TemplateFnReturnType;
 }
 
 export interface InfoButtonToggleEvent {