+++ packages/react/src/components/info-button/info-button.react-template.tsx
@@ -4,18 +4,20 @@ import * as React from "react";
 import { DsoInfoButton } from "../../components";
 import { ComponentImplementation } from "../../templates";
 
-export const reactInfoButton: ComponentImplementation<InfoButton> = {
+export const reactInfoButton: ComponentImplementation<InfoButton<React.JSX.Element>> = {
   component: "infoButton",
   implementation: "react",
   template: () =>
-    function infoButtonTemplate({ label, active, secondary, dsoToggle }) {
+    function infoButtonTemplate({ label, active, secondary, dsoToggle, children }) {
       return (
         <DsoInfoButton
           label={label}
           active={active}
           secondary={secondary}
           onDsoToggle={(e: CustomEvent) => dsoToggle?.(e.detail)}
-        ></DsoInfoButton>
+        >
+          {children && <div slot="toggletip">{children}</div>}
+        </DsoInfoButton>
       );
     },
 };