+++ storybook/src/components/info-button/info-button.core-template.ts
@@ -1,20 +1,21 @@
 import { InfoButton } from "dso-toolkit";
-import { html } from "lit-html";
+import { TemplateResult, html } from "lit-html";
 
 import { ComponentImplementation } from "../../templates";
 
-export const coreInfoButton: ComponentImplementation<InfoButton> = {
+export const coreInfoButton: ComponentImplementation<InfoButton<TemplateResult>> = {
   component: "infoButton",
   implementation: "core",
   template: () =>
-    function infoButtonTemplate({ label, active, secondary, dsoToggle }) {
+    function infoButtonTemplate({ label, active, secondary, dsoToggle, children }) {
       return html`
         <dso-info-button
           label=${label}
           ?active=${active}
           ?secondary=${secondary}
           @dsoToggle=${(e: CustomEvent) => dsoToggle?.(e.detail)}
-        ></dso-info-button>
+          >${children && html`<div slot="toggletip">${children}</div>`}</dso-info-button
+        >
       `;
     },
 };